app/src/main/kotlin/es/voghdev/chucknorrisjokes/ui/presenter/JokeByCategoryPresenter.kt
package es.voghdev.chucknorrisjokes.ui.presenter

import es.voghdev.chucknorrisjokes.app.ResLocator
import es.voghdev.chucknorrisjokes.app.coroutine
import es.voghdev.chucknorrisjokes.model.JokeCategory
import es.voghdev.chucknorrisjokes.repository.ChuckNorrisRepository

class JokeByCategoryPresenter(val context: ResLocator, val repository: ChuckNorrisRepository) :
        Presenter<JokeByCategoryPresenter.MVPView, JokeByCategoryPresenter.Navigator>() {

    override suspend fun initialize() {
        coroutine {
            repository.getJokeCategories()
        }
    }

    interface MVPView {
        fun fillCategoriesSpinner(categories: List<JokeCategory>)

    }

    interface Navigator {

    }
}
